requirements.txt
streamlit==1.35.0
pandas>=2.1
numpy>=1.26
scikit-learn>=1.5
pytest>=8.0
pandera==0.18.0
evidently==0.7.2
great-expectations==0.18.12
giskard-client==2.0.1
fairlearn==0.10.0
shap==0.45.0
captum==0.7.0
adversarial-robustness-toolbox==1.19.1
model-card-toolkit==0.7.3
reportlab==4.0.8
weasyprint==61.0
matplotlib>=3.9